% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_attachment.R
\name{extract_gwas_files}
\alias{extract_gwas_files}
\title{Extract one GWAS file path from each result object}
\usage{
extract_gwas_files(result_list)
}
\arguments{
\item{result_list}{List of attached result objects.}
}
\value{
Character vector of GWAS file paths.
}
\description{
Extract one GWAS file path from each result object
}
