% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_access.R
\name{mpgx_example_file}
\alias{mpgx_example_file}
\title{Locate a toy_data example file}
\usage{
mpgx_example_file(file)
}
\arguments{
\item{file}{File name under \code{inst/extdata}.}
}
\value{
Absolute file path.
}
\description{
Locate a toy_data example file
}
