% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_attachment.R
\name{mpgx_get_method_gwas_df}
\alias{mpgx_get_method_gwas_df}
\title{Extract parsed GWAS data frame from a method result object}
\usage{
mpgx_get_method_gwas_df(result_obj, label = "method")
}
\arguments{
\item{result_obj}{Method result object.}

\item{label}{Label used in error messages.}
}
\value{
Parsed GWAS data frame.
}
\description{
Extract parsed GWAS data frame from a method result object
}
