% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mpgx_run_bgenix_index}
\alias{mpgx_run_bgenix_index}
\title{Index a BGEN file with bgenix}
\usage{
mpgx_run_bgenix_index(bgen_file, bgenix_path = NULL, configs = NULL)
}
\arguments{
\item{bgen_file}{Path to BGEN file.}

\item{bgenix_path}{Optional path to \code{bgenix}.}

\item{configs}{Optional named list of executable overrides such as \code{bgenix.path}.}
}
\value{
Command metadata list.
}
\description{
Index a BGEN file with bgenix
}
